# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Submit a pending customer request
      description: Creates a pending request only. This does not take payment or confirm fulfillment by a provider.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Request accepted as pending, or the original result for an identical idempotent retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCreated"
        "400":
          description: Invalid request
        "401":
          description: Not signed in
        "409":
          description: The idempotency key was already used with a different payload
    get:
      operationId: getOrders
      tags: [orders]
      summary: List the signed-in customer's requests
      responses:
        "200":
          description: Requests owned by the signed-in customer
          content:
            application/json:
              schema:
                type: object
                required: [orders]
                properties:
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
        "401":
          description: Not signed in
  /orders/{id}:
    patch:
      operationId: cancelOrder
      tags: [orders]
      summary: Atomically cancel the signed-in customer's pending or accepted service booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerCancellationInput"
      responses:
        "200":
          description: Cancelled customer booking
          content:
            application/json:
              schema:
                type: object
                required: [order]
                properties:
                  order:
                    $ref: "#/components/schemas/Order"
        "400":
          description: Invalid cancellation or missing reason
        "401":
          description: Not signed in
        "409":
          description: Booking is not owned service work in a cancellable state, or another transition won
...
  /admin/providers:
    get:
      operationId: getAdminProviders
      tags: [admin]
      summary: List provider applications for review
      responses:
        "200":
          description: Non-demo provider applications
          content:
            application/json:
              schema:
                type: object
                required: [providers]
                properties:
                  providers:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminProviderProfile"
        "401":
          description: Not signed in
        "403":
          description: Admin role required
  /admin/providers/{id}:
    patch:
      operationId: reviewAdminProvider
      tags: [admin]
      summary: Approve, reject, or suspend a provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminProviderReview"
      responses:
        "200":
          description: Reviewed provider
          content:
            application/json:
              schema:
                type: object
                required: [provider]
                properties:
                  provider:
                    $ref: "#/components/schemas/AdminProviderProfile"
        "400":
          description: Invalid review or required reason missing
        "401":
          description: Not signed in
        "403":
          description: Admin role required
        "404":
          description: Application not found
  /demo/orders/{id}/respond:
    post:
      operationId: respondToDemoOrder
      tags: [demo]
      summary: Development-only creator-controlled response for a demo-provider request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderResponseInput"
      responses:
        "200":
          description: Clearly labelled test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoOrderResponse"
        "400":
          description: Invalid decision
        "401":
          description: Not signed in
        "403":
          description: Not the creator or not a demo request
        "404":
          description: Route disabled outside development
        "409":
          description: Demo request already answered
...
components:
  schemas:
...
    OrderKind:
      type: string
      enum: [service, product]
    OrderStatus:
      type: string
      enum: [pending, accepted, rejected, completed, cancelled]
    CancellationActor:
      type: string
      enum: [customer, provider]
    OrderItem:
      type: object
      additionalProperties: false
      required: [id, name, quantity]
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 200
        name:
          type: string
          minLength: 1
          maxLength: 200
        quantity:
          type: integer
          minimum: 1
          maximum: 1000
...
    ServiceBookingNeeds:
      type: object
      additionalProperties: false
      properties:
        approximateAreaM2:
          type: number
          exclusiveMinimum: 0
        shrubCount:
          type: integer
          minimum: 1
        standingTreeCount:
          type: integer
          minimum: 1
        stumpOrRootAreaCount:
          type: integer
          minimum: 1
        treeAccessDetails:
          type: string
          minLength: 1
          maxLength: 2000
    ServiceBooking:
      type: object
      additionalProperties: false
      required: [location, taskIds, preferredTime]
      properties:
        gardenName:
          type: string
          maxLength: 200
        location:
          type: string
          maxLength: 1000
        taskIds:
          type: array
          maxItems: 20
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 100
        notes:
          type: string
          maxLength: 5000
        needs:
          $ref: "#/components/schemas/ServiceBookingNeeds"
        preferredTime:
          type: string
          maxLength: 100
        photoPaths:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            pattern: "^/objects/uploads/[A-Za-z0-9_-]+/[A-Za-z0-9_-]+$"
...
    CreateOrderRequest:
      type: object
      additionalProperties: false
      required: [kind, items, idempotencyKey]
      properties:
        kind:
          $ref: "#/components/schemas/OrderKind"
        items:
          type: array
          minItems: 1
          maxItems: 50
          items:
            $ref: "#/components/schemas/OrderItem"
        details:
          type: string
          maxLength: 5000
        providerId:
          type: string
          minLength: 1
          maxLength: 200
        booking:
          $ref: "#/components/schemas/ServiceBooking"
        idempotencyKey:
          type: string
          minLength: 1
          maxLength: 200
    OrderCreated:
      type: object
      additionalProperties: false
      required: [id, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/OrderStatus"
        createdAt:
          type: string
          format: date-time
    Order:
      type: object
      additionalProperties: false
      required: [id, kind, items, status, responseReason, respondedAt, cancellationReason, cancelledAt, cancelledBy, completedAt, createdAt]
      properties:
        id:
          type: string
          format: uuid
        kind:
          $ref: "#/components/schemas/OrderKind"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        details:
          type: ["string", "null"]
        providerId:
          type: ["string", "null"]
        booking:
          oneOf:
            - $ref: "#/components/schemas/ServiceBooking"
            - type: "null"
        status:
          $ref: "#/components/schemas/OrderStatus"
        responseReason:
          type: ["string", "null"]
        respondedAt:
          type: ["string", "null"]
          format: date-time
        cancellationReason:
          type: ["string", "null"]
        cancelledAt:
          type: ["string", "null"]
          format: date-time
        cancelledBy:
          oneOf:
            - $ref: "#/components/schemas/CancellationActor"
            - type: "null"
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    ServicePrices:
      type: object
      additionalProperties:
        type: ["number", "null"]
        minimum: 0
    ProviderStatus:
      type: string
      enum: [draft, pending, approved, rejected, suspended]
    ProviderProfileEdit:
      type: object
      additionalProperties: false
      required: [shopName, serviceIds, description, earliestAvailableAt, servicePrices]
      properties:
        shopName:
          type: string
          minLength: 2
          maxLength: 150
        serviceIds:
          type: array
          minItems: 1
          maxItems: 50
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 100
        description:
          type: string
          minLength: 1
          maxLength: 5000
        earliestAvailableAt:
          type: ["string", "null"]
          format: date-time
        servicePrices:
          $ref: "#/components/schemas/ServicePrices"
    ProviderProfile:
      allOf:
        - $ref: "#/components/schemas/ProviderProfileEdit"
        - type: object
          required: [id, status, statusReason, submittedAt, reviewedAt, isDemo, createdAt, updatedAt]
          properties:
            id:
              type: string
            status:
              $ref: "#/components/schemas/ProviderStatus"
            statusReason:
              type: ["string", "null"]
            submittedAt:
              type: ["string", "null"]
              format: date-time
            reviewedAt:
              type: ["string", "null"]
              format: date-time
            isDemo:
              type: boolean
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    AdminProviderProfile:
      allOf:
        - $ref: "#/components/schemas/ProviderProfile"
        - type: object
          required: [applicantName]
          properties:
            applicantName:
              type: string
...
    OrderResponseInput:
      oneOf:
        - type: object
          additionalProperties: false
          required: [status]
          properties:
            status:
              type: string
              enum: [accepted, completed]
            reason:
              type: string
              maxLength: 2000
        - type: object
          additionalProperties: false
          required: [status, reason]
          properties:
            status:
              type: string
              const: rejected
            reason:
              type: string
              minLength: 1
              maxLength: 2000
        - type: object
          additionalProperties: false
          required: [status, reason]
          properties:
            status:
              type: string
              const: cancelled
            reason:
              type: string
              minLength: 1
              maxLength: 1000
    CustomerCancellationInput:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        status:
          type: string
          const: cancelled
        reason:
          type: string
          minLength: 1
          maxLength: 1000
    AdminProviderReview:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected, suspended]
        reason:
          type: string
          maxLength: 2000
    DemoOrderResponse:
      type: object
      required: [testMode, testStatus, order]
      properties:
        testMode:
          type: boolean
          const: true
        testStatus:
          type: string
        order:
          type: object
          required: [id, status, responseReason, respondedAt, cancellationReason, cancelledAt, cancelledBy, completedAt]
          properties:
            id:
              type: string
              format: uuid
            status:
              $ref: "#/components/schemas/OrderStatus"
            responseReason:
              type: ["string", "null"]
            respondedAt:
              type: ["string", "null"]
              format: date-time
            cancellationReason:
              type: ["string", "null"]
            cancelledAt:
              type: ["string", "null"]
              format: date-time
            cancelledBy:
              oneOf:
                - $ref: "#/components/schemas/CancellationActor"
                - type: "null"
            completedAt:
              type: ["string", "null"]
              format: date-time