# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CancellationActor } from './cancellationActor';
import type { OrderItem } from './orderItem';
import type { OrderKind } from './orderKind';
import type { OrderStatus } from './orderStatus';
import type { ServiceBooking } from './serviceBooking';

export interface Order {
  id: string;
  kind: OrderKind;
  items: OrderItem[];
  /** @nullable */
  details?: string | null;
  /** @nullable */
  providerId?: string | null;
  booking?: ServiceBooking | null;
  status: OrderStatus;
  /** @nullable */
  responseReason: string | null;
  /** @nullable */
  respondedAt: Date | null;
  /** @nullable */
  cancellationReason: string | null;
  /** @nullable */
  cancelledAt: Date | null;
  cancelledBy: CancellationActor | null;
  /** @nullable */
  completedAt: Date | null;
  createdAt: Date;
}