# lib/api-zod/src/generated/types/orderKind.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type OrderKind = typeof OrderKind[keyof typeof OrderKind];


export const OrderKind = {
  service: 'service',
  product: 'product',
} as const;