# lib/api-zod/src/generated/types/serviceBookingNeeds.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceBookingNeeds {
  /** @exclusiveMinimum 0 */
  approximateAreaM2?: number;
  /** @minimum 1 */
  shrubCount?: number;
  /** @minimum 1 */
  standingTreeCount?: number;
  /** @minimum 1 */
  stumpOrRootAreaCount?: number;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  treeAccessDetails?: string;
}